requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
pytest==7.4.2
httpx==0.24.1
